requirements.txt
fastapi
uvicorn[standard]
aiogram==2.25.1
python-dotenv
pydantic